requirements.txt
python-telegram-bot==21.9
  pymongo[srv]==4.10.1
  python-dotenv==1.0.1
  apscheduler==3.10.4
  